refactor: Order 와 관련된 Command 전용 응용 서비스를 생성한다.
